layouts added
layouts added, and useful functions

+++ hunt_or_be_hunted/src/main/java/project/hunt_or_be_hunted/Food_source.java
@@ -9,9 +9,9 @@ package project.hunt_or_be_hunted;
  * @author filip
  */
 public class Food_source extends Enviroment {
-    private float replenish_speed;
+    private int replenish_speed;
 
-    public Food_source(int x, int y, int capacity, float _replenish_speed) {
+    public Food_source(int x, int y, int capacity, int _replenish_speed) {
         super(x, y, capacity);
         replenish_speed = _replenish_speed;
     }
