+++ hunt_or_be_hunted/src/main/java/project/hunt_or_be_hunted/Water_source.java
@@ -9,9 +9,9 @@ package project.hunt_or_be_hunted;
  * @author filip
  */
 public class Water_source extends Enviroment {
-    private float replenish_speed;
+    private int replenish_speed;
 
-    public Water_source(int x, int y, int capacity, float _replenish_speed) {
+    public Water_source(int x, int y, int capacity, int _replenish_speed) {
         super(x, y, capacity);
         replenish_speed = _replenish_speed;
     }
